refactor: split gex subtab sections
