+++ web/components/regime/GexSubTab.tsx
@@ -1,64 +1,26 @@
 "use client";
 
-import { useMemo, useState } from "react";
 import { Activity, TrendingUp, TrendingDown } from "lucide-react";
 import {
   useGex,
-  type GexData,
-  type GexHistoryEntry,
   type GexLevel,
   type MqLevels,
   type SourceDelta,
-  type SourceDeltaEntry,
 } from "@/lib/regime/useGex";
 import { MarketState } from "@/lib/regime/useMarketHours";
 import InfoTooltip from "./InfoTooltip";
 import GexProfileChart from "./GexProfileChart";
 import { HistoryChart } from "./HistoryChart";
+import { ExpectedRangeBar } from "./gex/ExpectedRangeBar";
+import { GexHistoryTable } from "./gex/GexHistoryTable";
+import { MqLevelsPanel } from "./gex/MqLevelsPanel";
+import { biasColor, biasLabel, fmtGex, fmtPrice } from "./gex/format";
 import { formatPercent } from "./primitives/format";
 import { MetricCard, SourceBadge } from "./ui/MetricCard";
 
 type GexSubTabProps = {
   marketState?: MarketState;
 };
-
-/* ─── Helpers ─────────────────────────────────────────── */
-
-function fmtGex(v: number | null | undefined): string {
-  if (v == null) return "---";
-  const absVal = Math.abs(v);
-  if (absVal >= 1_000_000)
-    return `${v >= 0 ? "+" : ""}$${(v / 1_000_000).toFixed(1)}M`;
-  if (absVal >= 1_000) return `${v >= 0 ? "+" : ""}$${(v / 1_000).toFixed(1)}K`;
-  return `${v >= 0 ? "+" : ""}$${v.toFixed(0)}`;
-}
-
-function fmtPrice(v: number | null | undefined): string {
-  if (v == null) return "---";
-  return v.toLocaleString("en-US", {
-    minimumFractionDigits: 2,
-    maximumFractionDigits: 2,
-  });
-}
-
-function biasColor(direction: string): string {
-  switch (direction) {
-    case "BULL":
-      return "var(--signal-core)";
-    case "CAUTIOUS_BULL":
-      return "var(--signal-core)";
-    case "BEAR":
-      return "var(--fault)";
-    case "CAUTIOUS_BEAR":
-      return "var(--fault)";
-    default:
-      return "var(--neutral)";
-  }
-}
-
-function biasLabel(direction: string): string {
-  return direction.replace("_", " ");
-}
 
 /* ─── Spot freshness pill ─────────────────────────────── */
 
@@ -157,479 +119,6 @@ function LevelCard({
       <div className="gex-level-sub">
         {formatPercent(level.distance_pct)} &mdash; {fmtGex(level.gamma)} per $1
       </div>
-    </div>
-  );
-}
-
-/* ─── Expected Range Bar ─────────────────────────────── */
-
-function ExpectedRangeBar({ data }: { data: GexData }) {
-  const { expected_range, levels, spot } = data;
-  if (!expected_range.low || !expected_range.high) return null;
-
-  const low = expected_range.low;
-  const high = expected_range.high;
-  const flip = levels.gex_flip?.strike;
-  const magnet = levels.max_magnet?.strike;
-  const accel = levels.max_accelerator?.strike;
-
-  const allPoints = [low, high, spot];
-  if (flip) allPoints.push(flip);
-  if (magnet) allPoints.push(magnet);
-  if (accel) allPoints.push(accel);
-  const minVal = Math.min(...allPoints);
-  const maxVal = Math.max(...allPoints);
-  const range = maxVal - minVal || 1;
-  const pct = (v: number) => ((v - minVal) / range) * 100;
-
-  return (
-    <div className="gex-range-container">
-      <div className="gex-range-title">
-        EXPECTED RANGE &mdash; {data.data_date}
-      </div>
-      <div className="gex-range-bar">
-        <div
-          className="gex-range-fill"
-          style={{ left: `${pct(low)}%`, width: `${pct(high) - pct(low)}%` }}
-        />
-        {/* Markers */}
-        {flip && (
-          <div
-            className="gex-range-marker"
-            style={{ left: `${pct(flip)}%`, borderColor: "var(--warning)" }}
-            title={`GEX FLIP: ${fmtPrice(flip)}`}
-          />
-        )}
-        <div
-          className="gex-range-marker"
-          style={{ left: `${pct(spot)}%`, borderColor: "var(--signal-strong)" }}
-          title={`SPOT: ${fmtPrice(spot)}`}
-        />
-        {magnet && (
-          <div
-            className="gex-range-marker"
-            style={{
-              left: `${pct(magnet)}%`,
-              borderColor: "var(--signal-core)",
-            }}
-            title={`MAGNET: ${fmtPrice(magnet)}`}
-          />
-        )}
-      </div>
-      <div className="gex-range-labels">
-        <span>{fmtPrice(low)}</span>
-        {flip && (
-          <span style={{ left: `${pct(flip)}%`, color: "var(--warning)" }}>
-            {fmtPrice(flip)}
-          </span>
-        )}
-        <span style={{ marginLeft: "auto" }}>{fmtPrice(high)}</span>
-      </div>
-      <div className="gex-range-sublabels">
-        {accel && <span>MAX ACCEL</span>}
-        {flip && <span>GEX FLIP</span>}
-        <span>CLOSE</span>
-        {magnet && <span>MAX MAGNET</span>}
-      </div>
-    </div>
-  );
-}
-
-/* ─── History Table ──────────────────────────────────── */
-
-type GexSortCol =
-  | "date"
-  | "net_gex"
-  | "net_dex"
-  | "gex_flip"
-  | "spot"
-  | "atm_iv"
-  | "vol_pc"
-  | "bias";
-type SortDir = "asc" | "desc";
-
-function sortIndicator(
-  col: GexSortCol,
-  activeCol: GexSortCol | null,
-  dir: SortDir,
-): string {
-  if (col !== activeCol) return "";
-  return dir === "asc" ? " \u2191" : " \u2193";
-}
-
-/* ─── MenthorQ Levels Panel ─────────────────────────── */
-
-function MqLevelsPanel({
-  mq,
-  sourceDelta,
-}: {
-  mq: MqLevels;
-  sourceDelta: SourceDelta | null;
-}) {
-  const [expanded, setExpanded] = useState(false);
-
-  function deltaStyle(d: number | undefined): React.CSSProperties {
-    if (d == null) return {};
-    if (Math.abs(d) <= 2) return { color: "var(--signal-core)" };
-    if (Math.abs(d) <= 10) return { color: "var(--warning)" };
-    return { color: "var(--fault)" };
-  }
-
-  function fmtDelta(e: SourceDeltaEntry | undefined): React.ReactNode {
-    if (!e) return <span style={{ color: "var(--text-muted)" }}>—</span>;
-    const sign = e.delta > 0 ? "+" : "";
-    return (
-      <span style={deltaStyle(e.delta)}>
-        {sign}
-        {e.delta.toFixed(1)} &nbsp;
-        <span style={{ color: "var(--signal-core)", fontSize: 9 }}>
-          {e.uw.toFixed(0)}
-        </span>
-        <span style={{ color: "var(--text-muted)", fontSize: 9 }}> vs </span>
-        <span style={{ color: "#85b7eb", fontSize: 9 }}>{e.mq.toFixed(0)}</span>
-      </span>
-    );
-  }
-
-  return (
-    <div className="gex-history-section">
-      <button className="gex-mq-toggle" onClick={() => setExpanded(!expanded)}>
-        MenthorQ Key Levels
-        {mq.source_date && (
-          <span
-            style={{
-              fontFamily: "var(--font-mono)",
-              fontSize: 10,
-              color: "var(--text-muted)",
-              marginLeft: 8,
-            }}
-          >
-            {mq.source_date}
-          </span>
-        )}
-        <SourceBadge source="mq" /> {expanded ? "▲" : "▼"}
-      </button>
-      {expanded && (
-        <div
-          style={{
-            display: "grid",
-            gridTemplateColumns: "1fr 1fr",
-            gap: 16,
-            marginTop: 10,
-          }}
-        >
-          {/* MQ Level Values */}
-          <div>
-            <div
-              style={{
-                fontSize: 10,
-                letterSpacing: "0.1em",
-                textTransform: "uppercase",
-                color: "var(--text-muted)",
-                marginBottom: 8,
-              }}
-            >
-              Levels
-            </div>
-            {[
-              { label: "HVL (flip)", val: mq.hvl },
-              { label: "Call Resistance (all)", val: mq.call_resistance_all },
-              { label: "Call Resistance (0DTE)", val: mq.call_resistance_0dte },
-              { label: "Put Support (all)", val: mq.put_support_all },
-              { label: "Put Support (0DTE)", val: mq.put_support_0dte },
-              { label: "Expected High", val: mq.expected_high },
-              { label: "Expected Low", val: mq.expected_low },
-            ].map(({ label, val }) => (
-              <div
-                key={label}
-                style={{
-                  display: "flex",
-                  justifyContent: "space-between",
-                  fontSize: 11,
-                  marginBottom: 5,
-                  fontFamily: "var(--font-mono)",
-                }}
-              >
-                <span style={{ color: "var(--text-secondary)", fontSize: 10 }}>
-                  {label}
-                </span>
-                <span style={{ color: "#85b7eb", fontWeight: 500 }}>
-                  {val != null ? fmtPrice(val) : "—"}
-                </span>
-              </div>
-            ))}
-            {mq.top_gex_strikes.length > 0 && (
-              <div style={{ marginTop: 8 }}>
-                <div
-                  style={{
-                    fontSize: 10,
-                    color: "var(--text-muted)",
-                    marginBottom: 4,
-                  }}
-                >
-                  Top GEX Strikes
-                </div>
-                <div style={{ display: "flex", gap: 6, flexWrap: "wrap" }}>
-                  {mq.top_gex_strikes.map((s) => (
-                    <span
-                      key={s}
-                      style={{
-                        background: "rgba(56,138,221,0.12)",
-                        color: "#85b7eb",
-                        border: "0.5px solid rgba(56,138,221,0.3)",
-                        fontSize: 10,
-                        padding: "1px 6px",
-                        borderRadius: 2,
-                        fontFamily: "var(--font-mono)",
-                      }}
-                    >
-                      {fmtPrice(s)}
-                    </span>
-                  ))}
-                </div>
-              </div>
-            )}
-          </div>
-
-          {/* Source Delta */}
-          <div>
-            <div
-              style={{
-                fontSize: 10,
-                letterSpacing: "0.1em",
-                textTransform: "uppercase",
-                color: "var(--text-muted)",
-                marginBottom: 8,
-              }}
-            >
-              UW vs MQ Delta &nbsp;
-              <span
-                style={{
-                  color: "var(--text-muted)",
-                  fontStyle: "italic",
-                  textTransform: "none",
-                }}
-              >
-                (+= UW higher)
-              </span>
-            </div>
-            {sourceDelta ? (
-              [
-                { label: "Flip vs HVL", entry: sourceDelta.flip_vs_hvl },
-                {
-                  label: "Put wall vs support (all)",
-                  entry: sourceDelta.put_wall_vs_support_all,
-                },
-                {
-                  label: "Put wall vs support (0DTE)",
-                  entry: sourceDelta.put_wall_vs_support_0dte,
-                },
-                {
-                  label: "Call wall vs resist (all)",
-                  entry: sourceDelta.call_wall_vs_resistance_all,
-                },
-                {
-                  label: "Call wall vs resist (0DTE)",
-                  entry: sourceDelta.call_wall_vs_resistance_0dte,
-                },
-              ].map(({ label, entry }) => (
-                <div
-                  key={label}
-                  style={{
-                    display: "flex",
-                    justifyContent: "space-between",
-                    fontSize: 11,
-                    marginBottom: 5,
-                    fontFamily: "var(--font-mono)",
-                  }}
-                >
-                  <span
-                    style={{ color: "var(--text-secondary)", fontSize: 10 }}
-                  >
-                    {label}
-                  </span>
-                  {fmtDelta(entry)}
-                </div>
-              ))
-            ) : (
-              <div style={{ color: "var(--text-muted)", fontSize: 11 }}>
-                No delta data
-              </div>
-            )}
-            {/* IV comparison */}
-            {(mq.iv30d != null || mq.hv30 != null) && (
-              <div style={{ marginTop: 12 }}>
-                <div
-                  style={{
-                    fontSize: 10,
-                    color: "var(--text-muted)",
-                    marginBottom: 4,
-                  }}
-                >
-                  Volatility (MQ)
-                </div>
-                {mq.iv30d != null && (
-                  <div
-                    style={{
-                      display: "flex",
-                      justifyContent: "space-between",
-                      fontSize: 11,
-                      marginBottom: 4,
-                      fontFamily: "var(--font-mono)",
-                    }}
-                  >
-                    <span
-                      style={{ color: "var(--text-secondary)", fontSize: 10 }}
-                    >
-                      IV 30D
-                    </span>
-                    <span style={{ color: "#85b7eb", fontWeight: 500 }}>
-                      {(mq.iv30d * 100).toFixed(2)}%
-                    </span>
-                  </div>
-                )}
-                {mq.hv30 != null && (
-                  <div
-                    style={{
-                      display: "flex",
-                      justifyContent: "space-between",
-                      fontSize: 11,
-                      fontFamily: "var(--font-mono)",
-                    }}
-                  >
-                    <span
-                      style={{ color: "var(--text-secondary)", fontSize: 10 }}
-                    >
-                      HV 30D
-                    </span>
-                    <span
-                      style={{
-                        color: "var(--text-secondary)",
-                        fontWeight: 500,
-                      }}
-                    >
-                      {(mq.hv30 * 100).toFixed(2)}%
-                    </span>
-                  </div>
-                )}
-              </div>
-            )}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function GexHistoryTable({ history }: { history: GexHistoryEntry[] }) {
-  const [sortCol, setSortCol] = useState<GexSortCol | null>(null);
-  const [sortDir, setSortDir] = useState<SortDir>("desc");
-  const [expanded, setExpanded] = useState(false);
-
-  function handleSort(col: GexSortCol) {
-    if (sortCol === col) {
-      if (sortDir === "desc") setSortDir("asc");
-      else {
-        setSortCol(null);
-        setSortDir("desc");
-      }
-    } else {
-      setSortCol(col);
-      setSortDir("desc");
-    }
-  }
-
-  const sorted = useMemo(() => {
-    if (!sortCol) return history;
-    return [...history].sort((a, b) => {
-      const av = sortCol === "date" ? a.date : (a[sortCol] ?? -Infinity);
-      const bv = sortCol === "date" ? b.date : (b[sortCol] ?? -Infinity);
-      if (av < bv) return sortDir === "asc" ? -1 : 1;
-      if (av > bv) return sortDir === "asc" ? 1 : -1;
-      return 0;
-    });
-  }, [history, sortCol, sortDir]);
-
-  if (!history.length) return null;
-
-  const cols: { key: GexSortCol; label: string; align: string }[] = [
-    { key: "date", label: "Date", align: "left" },
-    { key: "spot", label: "Spot", align: "right" },
-    { key: "gex_flip", label: "GEX Flip", align: "right" },
-    { key: "net_gex", label: "Net GEX", align: "right" },
-    { key: "net_dex", label: "Net DEX", align: "right" },
-    { key: "atm_iv", label: "IV 30D", align: "right" },
-    { key: "vol_pc", label: "Vol P/C", align: "right" },
-    { key: "bias", label: "Bias", align: "center" },
-  ];
-
-  return (
-    <div className="gex-history-section">
-      <button
-        className="gex-history-toggle"
-        onClick={() => setExpanded(!expanded)}
-      >
-        History ({history.length} sessions) {expanded ? "\u25B2" : "\u25BC"}
-      </button>
-      {expanded && (
-        <div className="gex-history-table-wrap">
-          <table className="gex-history-table">
-            <thead>
-              <tr>
-                {cols.map((c) => (
-                  <th
-                    key={c.key}
-                    className={`text-${c.align}`}
-                    onClick={() => handleSort(c.key)}
-                    style={{ cursor: "pointer", userSelect: "none" }}
-                  >
-                    {c.label}
-                    {sortIndicator(c.key, sortCol, sortDir)}
-                  </th>
-                ))}
-              </tr>
-            </thead>
-            <tbody>
-              {sorted.map((row) => (
-                <tr key={row.date}>
-                  <td>{row.date}</td>
-                  <td className="text-right">{fmtPrice(row.spot)}</td>
-                  <td className="text-right">{fmtPrice(row.gex_flip)}</td>
-                  <td
-                    className="text-right"
-                    style={{
-                      color:
-                        row.net_gex >= 0
-                          ? "var(--signal-core)"
-                          : "var(--fault)",
-                    }}
-                  >
-                    {fmtGex(row.net_gex)}
-                  </td>
-                  <td className="text-right">{fmtGex(row.net_dex)}</td>
-                  <td className="text-right">
-                    {row.atm_iv != null ? `${row.atm_iv.toFixed(1)}%` : "---"}
-                  </td>
-                  <td className="text-right">
-                    {row.vol_pc != null ? row.vol_pc.toFixed(2) : "---"}
-                  </td>
-                  <td className="text-center">
-                    <span
-                      style={{
-                        color: biasColor(row.bias || "NEUTRAL"),
-                        fontWeight: 600,
-                        fontSize: 10,
-                      }}
-                    >
-                      {biasLabel(row.bias || "NEUTRAL")}
-                    </span>
-                  </td>
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        </div>
-      )}
     </div>
   );
 }
